+++ web_search_neo/__init__.py
@@ -1,3 +1,3 @@
 """Web Search Neo package."""
 
-__version__ = '1.16.2'
+__version__ = '1.16.3'